requirements.txt
pydantic>=2.7.3
pydantic[email]>=2.7.3
email-validator>=2.1.0
weasyprint>=64.0
Pillow>=11.0.0
pybars3>=0.9.7
pdf2image>=1.17.0
python-dateutil>=2.8.2
openai>=1.0.0
azure-identity>=1.15.0
python-dotenv>=1.0.0
crewai>=0.1.0
crewai-tools>=0.1.0
langchain>=0.1.0
langchain-community>=0.1.0
beautifulsoup4>=4.12.0
requests>=2.31.0 